Move functions from `utils` to `docs.common`
I put it in `utils` originally, but on second thoughts I'd rather keep
this kind of gnarly code which is only used for docs generation out of
the main codebase.
